+++ src/main/java/com/lescastcodeurs/bot/ShowNoteCategory.java
@@ -3,39 +3,45 @@ package com.lescastcodeurs.bot;
 import static java.util.Objects.requireNonNull;
 
 import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
-/** Les cast codeurs podcast categories. */
+/** Les cast codeurs podcast categories (by order of appearance in the show notes). */
 public enum ShowNoteCategory {
-  NEWS("Non catégorisées", "news", "nouvelles", "nouvelle"),
-  LANGUAGES("Langages", "langages", "langage", "languages", "language", "langs", "lang"),
-  LIBRARIES("Librairies", "libraries", "librairies", "librairie", "library", "libs", "lib"),
-  INFRASTRUCTURE("Infrastructure", "infrastructure", "infra"),
-  CLOUD("Cloud", "cloud"),
-  WEB("Web", "web"),
-  DATA("Data", "data"),
-  TOOLING("Outillage", "outillage", "tooling"),
-  ARCHITECTURE("Architecture", "architecture", "architectures", "archi", "arch"),
+  LANGUAGES("Langages", "lang", "langage", "langages", "langs", "language", "languages"),
+  LIBRARIES("Librairies", "lib", "libs", "libraries", "librairies", "librairie", "library"),
+  INFRASTRUCTURE("Infrastructure", "infra", "infrastructure"),
+  CLOUD("Cloud", "cloud", "iaas"),
+  WEB("Web", "web", "www"),
+  DATA("Data", "data", "db"),
+  TOOLING("Outillage", "outil", "outils", "tool", "tools", "outillage", "tooling"),
+  ARCHITECTURE("Architecture", "archi", "arch", "architecture", "architectures"),
   METHODOLOGIES(
-      "Méthodologies", "methodologies", "methodologie", "methodology", "methode", "methodo"),
-  SECURITY("Sécurité", "securite", "security", "secure", "secu", "sec"),
+      "Méthodologies",
+      "methodo",
+      "metodo",
+      "methode",
+      "methodologie",
+      "methodologies",
+      "methodology"),
+  SECURITY("Sécurité", "secu", "sec", "securite", "security", "secure"),
   SOCIETY(
       "Loi, société et organisation",
       "loi",
       "law",
       "societe",
       "society",
-      "organisation",
-      "organization",
+      "org",
       "orga",
-      "org"),
-  TOOL_OF_THE_EPISODE("Outils de l’épisode", "outils", "outil", "tools", "tool"),
-  BEGINNERS("Rubrique débutant", "debutants", "debutant", "beginners", "beginner"),
-  CONFERENCE("Conférences", "conferences", "conference", "conf");
+      "organisation",
+      "organization"),
+  TOOL_OF_THE_EPISODE("Outils de l’épisode", "outil-ep", "outil-episode"),
+  BEGINNERS("Rubrique débutant", "debutant", "debutants", "beginner", "beginners"),
+  CONFERENCES("Conférences", "conf", "conferences", "conference"),
+  NEWS("Non catégorisé", "news", "nouvelles", "nouvelle");
 
   private final String description;
-  private final Set<String> labels;
+  private final List<String> labels;
 
   /**
    * Returns the category matching the given non-null label, or {@link #NEWS} if the label does not
@@ -48,7 +54,7 @@ public enum ShowNoteCategory {
       String normalized = StringUtils.normalize(label);
 
       for (ShowNoteCategory category : values()) {
-        if (category.labels.contains(normalized)) {
+        if (category.labels().contains(normalized)) {
           return Optional.of(category);
         }
       }
@@ -59,14 +65,27 @@ public enum ShowNoteCategory {
 
   ShowNoteCategory(String description, String... labels) {
     this.description = requireNonNull(description);
-    this.labels = Set.copyOf(Arrays.stream(labels).map(StringUtils::normalize).toList());
+    this.labels = List.copyOf(Arrays.stream(labels).map(StringUtils::normalize).toList());
+
+    if (this.labels.size() < 2) {
+      throw new IllegalArgumentException(
+          "at least two labels are required (formatting help is easier with this constraint)");
+    }
   }
 
   public String description() {
     return description;
   }
 
-  public Set<String> getLabels() {
+  public String mainLabel() {
+    return labels.get(0);
+  }
+
+  public List<String> alternateLabels() {
+    return labels.subList(1, labels.size());
+  }
+
+  public List<String> labels() {
     return labels;
   }
 }